backend/vision-analysis/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
google-generativeai==0.7.2
Pillow==10.3.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
pydantic==2.7.1
numpy==1.26.4
aiofiles==23.2.1
